added database table seeder

+++ app/database/migrations/2014_11_18_201753_create_Materials_table.php
@@ -16,7 +16,7 @@ class CreateMaterialsTable extends Migration {
 		{
 			$table->increments('id');
 			$table->string('name');
-			$table->float('priceId');
+			$table->integer('priceId');
 			$table->timestamps();
 		});
 	}
